[FE] refactor : SnackbarProvider, Snackbar 컴포넌트와 통합
