+++ frontend/src/common/components/Snackbar/index.tsx
@@ -1,97 +1,30 @@
-import React, { useContext, useState } from 'react';
+import { snackbarStackWithActive } from 'common/recoil/snackbar';
 
-import { IconProp } from '@fortawesome/fontawesome-svg-core';
-import { faBell } from '@fortawesome/free-regular-svg-icons';
-import { faClose } from '@fortawesome/free-solid-svg-icons';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+import { useRecoilValue } from 'recoil';
 
-import cn from 'classnames';
+import useSnackbar from 'common/hooks/useSnackbar';
 
-import { TransitionDiv, Text } from 'common/components';
+import SnackbarComponent from 'common/components/Snackbar/component';
 
-import styles from './styles.module.scss';
+function SnackbarProvider() {
+  const snackbarActivityStack = useRecoilValue(snackbarStackWithActive);
+  const snackbar = useSnackbar();
 
-import { UserAgentContext } from 'common/contexts/UserAgent';
-
-export interface SnackbarProps {
-  icon?: IconProp;
-  theme?: 'primary' | 'success' | 'warning' | 'danger';
-  title: string;
-  description?: string;
-  duration?: number;
-  onDisappear?: React.AnimationEventHandler<HTMLDivElement>;
-}
-
-function Snackbar({
-  icon = faBell,
-  theme = 'primary',
-  title,
-  description,
-  duration = 3000,
-  onDisappear,
-}: SnackbarProps) {
-  const { isPC } = useContext(UserAgentContext);
-  const [isVisible, setVisible] = useState(true);
-
-  const progressDuration = `${(duration / 1000).toFixed(2)}s`;
-
-  const onProgressEnd = ({
-    type,
-    target,
-    currentTarget,
-  }: React.MouseEvent | React.AnimationEvent) => {
-    if (type === 'animationend' && target !== currentTarget) return;
-
-    setVisible(false);
-  };
-
-  const onDisappearContainer = (event: React.AnimationEvent<HTMLDivElement>) => {
-    onDisappear && onDisappear(event);
+  const handleRemoveSnackbar = (snackbarId: number) => () => {
+    snackbar.remove(snackbarId);
   };
 
   return (
-    <TransitionDiv
-      className={cn(styles.componentSnackbar, styles[theme])}
-      all="drop"
-      direction={isPC ? 'right' : 'up'}
-      duration={500}
-      onClick={onProgressEnd}
-      onDisappear={onDisappearContainer}
-      isVisible={isVisible}
-    >
-      <div className={styles.content}>
-        <FontAwesomeIcon className={styles.icon} icon={icon} />
-
-        <div className={styles.text}>
-          <Text className={styles.title} size={18} weight="bold">
-            {title}
-          </Text>
-          {description && (
-            <Text className={styles.description} size={14}>
-              {description}
-            </Text>
-          )}
-        </div>
-
-        <FontAwesomeIcon className={styles.close} icon={faClose} />
-      </div>
-
-      <div className={styles.progress}>
-        <div
-          className={styles.percent}
-          onAnimationEnd={onProgressEnd}
-          style={{ animationDuration: progressDuration }}
-        ></div>
-      </div>
-    </TransitionDiv>
+    <div id="snackbar-container">
+      {snackbarActivityStack.map(({ key, ...rest }) => (
+        <SnackbarComponent key={key} {...rest} onDisappear={handleRemoveSnackbar(key)} />
+      ))}
+    </div>
   );
 }
 
-Snackbar.defaultType = {
-  icon: faBell,
-  theme: 'primary',
-  title: '스낵바 타이틀이 비어있습니다',
-  duration: 3000,
-};
+const Snackbar = Object.assign(SnackbarComponent, {
+  Provider: SnackbarProvider,
+});
 
 export default Snackbar;
