+++ frontend/src/common/components/index.ts
@@ -10,7 +10,6 @@ export { default as ErrorBoundary } from './ErrorBoundary';
 export { default as TransitionDiv } from './TransitionDiv';
 export { default as Skeleton } from './Skeleton';
 export { default as Snackbar } from './Snackbar';
-export { default as SnackbarProvider } from './SnackbarProvider';
 export { default as PopupBox } from './PopupBox';
 export { default as Modal } from './Modal';
 export { default as FlexContainer } from './FlexContainer';
